+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-D7YdaSci.js
@@ -1,8 +1,0 @@
-import{c as r,s as e}from"./flowDiagram-HODETNUW-CVP4UfrA.js";import{_ as a}from"./mermaid.core-Dlq0a1VW.js";import"./chunk-5VM5RSS4-vcnngCGg.js";import"./chunk-XXDRQBXY-BmAiBRmV.js";import"./chunk-POPQ4Y6H-CpThFIui.js";import"./chunk-F27PBJKO-4QFkUBmV.js";import"./channel-DSKgAd8E.js";import"./index-C8XQ0fh1.js";var o=a(t=>`${e(t)}
-  .swimlane.cluster rect {
-    stroke: ${t.clusterBorder} !important;
-  }
-  [data-look="neo"].cluster rect {
-    filter: none;
-  }
-`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};